+++ modules/data_structures/builtin/cpp/Gateway.cpp
@@ -34,6 +34,7 @@
 #include "isfieldBuiltin.hpp"
 #include "namedargs2cellBuiltin.hpp"
 #include "getfieldBuiltin.hpp"
+#include "rmfieldBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -50,6 +51,7 @@ static const nlsGateway gateway[] = {
     { "isfield", Nelson::DataStructuresGateway::isfieldBuiltin, 1, 2 },
     { "namedargs2cell", Nelson::DataStructuresGateway::namedargs2cellBuiltin, 1, 1 },
     { "getfield", Nelson::DataStructuresGateway::getfieldBuiltin, 1, 2 },
+    { "rmfield", Nelson::DataStructuresGateway::rmfieldBuiltin, 1, 2 },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)